requirements.txt
discord.py==2.3.2
pydantic==2.5.2
pydantic-settings==2.1.0
pyautogui==0.9.54
psutil==5.9.6
